requirements.txt
python-dotenv
openai==1.14.2
langchain==0.1.13
langchain-community==0.0.29
langchain-core==0.1.33
langchain-openai==0.1.1
pydantic==1.10.13
urllib3==1.26.6